requirements.txt
django
pygments
djangorestframework
markdown
django-filter
django-crispy-forms
django-guardian
